pyproject.toml
[project]
name = "eris-rag"
version = "0.1.0"
description = "Character personality RAG enhancement server for AstrBot"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "chromadb>=0.5.0",
    "jieba>=0.42.1",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "sentence-transformers>=3.0.0",
    "FlagEmbedding>=1.2.0",
    "torch>=2.0.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
]
